Login/logout done
I added a login/logout controllers and views. In MY_Controller.php, I modified the render() function so it could determine whether or not the user is logged in or not and display the correct menu. Made so the log in page redirects the user to login page, if the incorrect information page was entered or the login link was pressed.